+++ src/app/locales/en/components/support-space-form.js
@@ -11,11 +11,6 @@ export default {
     size: {
       name: 'Size',
     },
-    importEnabled: {
-      name: 'Import storage data',
-      tipEnabled: 'You have chosen an import-enabled storage. The existing data will be imported into the space.',
-      tipDisabled: 'You must support the space with an import-enabled storage to unlock import settings.',
-    },
   },
   submitButton: 'Support space',
 };